Makes NavigationController::max_restored_page_id_ an int32 to match usage of it else where (TabContents, SiteInstance...)
BUG=none
TEST=none

Review URL: http://codereview.chromium.org/3173050

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@57591 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 61398156d8c68470a29914d77191f16df69b4c84

+++ chrome/browser/tab_contents/navigation_controller.cc
@@ -1110,7 +1110,7 @@ void NavigationController::FinishRestore(int selected_index,
   DCHECK(selected_index >= 0 && selected_index < entry_count());
   ConfigureEntriesForRestore(&entries_, from_last_session);
 
-  set_max_restored_page_id(entry_count());
+  set_max_restored_page_id(static_cast<int32>(entry_count()));
 
   last_committed_entry_index_ = selected_index;
 }
